# shared/openapi/leyo-swimming-v1.yaml
openapi: 3.0.3
info:
  title: leyoSwimming API
  description: 游泳私教预约平台后端 API 契约（登录注册 + 套餐购买订单退款批次）
  version: 1.1.0
  contact:
    name: leyoSwimming Team
...
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /api/admin/auth/login:
    post:
      tags:
        - Admin Auth
      summary: 管理员账号密码登录
      description: Web 管理后台登录接口，返回管理员 JWT。
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/auth/logout:
    post:
      tags:
        - Admin Auth
      summary: 管理员退出登录
      description: 失效当前管理员会话 token。
      operationId: adminLogout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/coach/list:
    post:
      tags:
        - Coach
      summary: 教练列表查询
      description: 游客可访问，支持分页。
      operationId: listCoaches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/coach/detail:
    post:
      tags:
        - Coach
      summary: 教练详情查询
      description: 游客可访问。错误码：COACH_NOT_FOUND
      operationId: getCoachDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/order/detail:
    post:
      tags:
        - Order
      summary: 订单详情
      description: 查询订单详情，返回订单、套餐、教练、支付、退款信息。错误码：ORDER_ACCESS_DENIED, ORDER_NOT_FOUND
      operationId: getOrderDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/order/list:
    post:
      tags:
        - Order
      summary: 订单列表
      description: 查询当前学员订单列表。
      operationId: listOrders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/order/list:
    post:
      tags:
        - Admin Order
      summary: 管理员查看订单列表
      description: 后台管理订单分页查询。错误码：FORBIDDEN
      operationId: listAdminOrders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseOrderListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/order/detail:
    post:
      tags:
        - Admin Order
      summary: 管理员查看订单详情
      description: 查询订单详情及关联记录。错误码：FORBIDDEN, ORDER_NOT_FOUND
      operationId: getAdminOrderDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseOrderDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    ApiResponseAdminLoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminLoginResponse'
...
    AdminLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
          description: 管理员会话 token（JWT），有效期 24 小时
        expiresIn:
          type: integer
          example: 86400
          description: token 有效期（秒）
        admin:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            name:
              type: string
            role:
              type: string
...
    PaginationRequest:
      type: object
      properties:
        page:
          type: integer
          default: 1
          minimum: 1
        pageSize:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
    CoachListRequest:
      $ref: '#/components/schemas/PaginationRequest'
    CoachListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: integer
        avatar:
          type: string
        rating:
          type: number
        yearsOfTeaching:
          type: integer
        teachingStrokes:
          type: array
          items:
            type: string
        realTimeStatus:
          type: string
    CoachListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CoachListItem'
    ApiResponseCoachListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachListPageResponse'
    CoachDetailRequest:
      type: object
      required:
        - coachId
      properties:
        coachId:
          type: integer
    CoachDetailContact:
      type: object
      properties:
        phone:
          type: string
        wechatQrCode:
          type: string
    CoachDetailCertificate:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
    CoachDetailPackage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        hours:
          type: integer
    CoachReview:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: number
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
    AvailableTime:
      type: object
      properties:
        date:
          type: string
          format: date
        slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
    CoachDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: integer
        age:
          type: integer
        status:
          type: integer
        avatar:
          type: string
        rating:
          type: number
        yearsOfTeaching:
          type: integer
        totalStudents:
          type: integer
        totalHours:
          type: integer
        teachingStrokes:
          type: array
          items:
            type: string
        bio:
          type: string
        referencePrice:
          type: integer
        contact:
          $ref: '#/components/schemas/CoachDetailContact'
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/CoachDetailCertificate'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/CoachDetailPackage'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/CoachReview'
        availableTimes:
          type: array
          items:
            $ref: '#/components/schemas/AvailableTime'
        realTimeStatus:
          type: string
    ApiResponseCoachDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachDetailResponse'
...
    OrderDetailRequest:
      type: object
      required:
        - orderId
      properties:
        orderId:
          type: integer
    OrderDetailPackage:
      type: object
      properties:
        packageId:
          type: integer
        packageName:
          type: string
    OrderDetailCoach:
      type: object
      properties:
        coachId:
          type: integer
        coachName:
          type: string
    OrderDetailPayment:
      type: object
      properties:
        paymentId:
          type: integer
        channelTradeNo:
          type: string
        status:
          type: string
    OrderDetailRefund:
      type: object
      properties:
        refundOrderId:
          type: integer
        status:
          type: string
        reason:
          type: string
    OrderDetailResponse:
      type: object
      properties:
        orderId:
          type: integer
        type:
          type: string
          enum: [purchase, refund]
        status:
          type: string
        amount:
          type: integer
        paidAt:
          type: string
          format: date-time
        packageId:
          type: integer
        package:
          $ref: '#/components/schemas/OrderDetailPackage'
        coach:
          $ref: '#/components/schemas/OrderDetailCoach'
        payment:
          $ref: '#/components/schemas/OrderDetailPayment'
        refund:
          $ref: '#/components/schemas/OrderDetailRefund'
    ApiResponseOrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderDetailResponse'
    OrderListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            status:
              type: string
    OrderListItem:
      type: object
      properties:
        orderId:
          type: integer
        type:
          type: string
        status:
          type: string
        statusName:
          type: string
        amount:
          type: integer
        packageMode:
          type: string
        packageModeLabel:
          type: string
        packageName:
          type: string
        coachName:
          type: string
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        refundReason:
          type: string
        remainingSeconds:
          type: integer
    OrderListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderListItem'
    ApiResponseOrderListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderListPageResponse'
...
    AdminApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
...
    AdminOrderListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            type:
              type: string
              enum: [purchase, refund]
            status:
              type: string
            coachId:
              type: integer
            userId:
              type: integer
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
    AdminOrderListItem:
      type: object
      properties:
        orderId:
          type: integer
        orderNo:
          type: string
        type:
          type: string
        status:
          type: string
        userId:
          type: integer
        userName:
          type: string
        coachId:
          type: integer
        coachName:
          type: string
        packageId:
          type: integer
        packageName:
          type: string
        amount:
          type: integer
        paidAmount:
          type: integer
        paymentMethod:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminOrderListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderListItem'
    AdminApiResponseOrderListResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminOrderListPageResponse'
    AdminOrderDetailRequest:
      type: object
      required:
        - orderId
      properties:
        orderId:
          type: integer
    AdminOrderDetailPackage:
      type: object
      properties:
        packageMode:
          type: string
        totalHours:
          type: integer
        consumedCount:
          type: integer
        paidAmount:
          type: integer
        originalPrice:
          type: integer
        refundEnabled:
          type: boolean
        refundRatio:
          type: number
        refundValidDays:
          type: integer
    AdminOrderDetailResponse:
      type: object
      properties:
        orderId:
          type: integer
        orderNo:
          type: string
        type:
          type: string
        status:
          type: string
        userId:
          type: integer
        userName:
          type: string
        coachId:
          type: integer
        coachName:
          type: string
        packageId:
          type: integer
        package:
          $ref: '#/components/schemas/AdminOrderDetailPackage'
        amount:
          type: integer
        paidAmount:
          type: integer
        paymentMethod:
          type: string
        channelTradeNo:
          type: string
        purchaseOrderId:
          type: integer
        reason:
          type: string
        rejectedReason:
          type: string
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        refundedAt:
          type: string
          format: date-time
        refundRecords:
          type: array
          items:
            type: object
        payment:
          type: object
        auditLog:
          type: array
          items:
            type: object
    AdminApiResponseOrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminOrderDetailResponse'
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100001
            message: 请求参数错误
            data: null
    Unauthorized:
      description: 未授权或 token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 200001
            message: 登录已过期，请重新登录
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100002
            message: 资源不存在
            data: null
    Forbidden:
      description: 权限禁止或账号被禁用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 300001
            message: 账号已被禁用，请联系超级管理员
            data: null
...